Move endpoints logic from main to controllers
