+++ src/server/__init__.py
@@ -1,3 +1,3 @@
 from .constants import PATHS, QUERIES
-from .main import app
 from .responses import Responses
+from .main import app
