+++ src/server/stock_market/stock_market_controller.py
@@ -1,0 +1,18 @@
+from flask import Blueprint
+
+class StockMarketController:
+    """Stock Market Controller class."""
+
+    blueprint = Blueprint('stock', __name__, url_prefix='/stock')
+
+    @staticmethod
+    @blueprint.route("/chart")
+    def chart() -> str:
+        """Chart data retrieval endpoint."""
+        return "chart"
+
+    @staticmethod
+    @blueprint.route("/future_value")
+    def future_value() -> str:
+        """Model data estimation endpoint."""
+        return "future_value"
