+++ src/server/stock_market/stock_market_service.py
@@ -1,0 +1,2 @@
+class StockMarketService:
+    """Stock Market Service class."""
